fix @namespace
I've found out that not changing the default value of `@namespace` might
not be ideal.  Change it to a proper value and add `@homepageURL`, while
we're here.

Changing `@namespace` like this could be problematic -- users might
recieve it as a separate script, but it shouldn't be a big problem, they
can just delete the old version.  Changing the `@namespace` doesn't seem
to affect the stored values.