comment on storage keys
When I implemented v1, the script was storing two kinds of values for
each playlist: the date and the video ID.  I kept adding values, because
it was a simple way of storing them.  But then I realized that it would
have been better to store a single object with all the necessary bits of
data for each playlist -- see infoStorageKey().  Unfortunately, at that
point, the script had already been published and used by a non-zero
number of people.  Even though the number is very small, it is still a
good exercise in keeping backwards compatibility.

Add comments clarifying that the stupid way of storing and loading the
dates is for backwards compatibility.