clean up line breaks in channelName
For some reasons, some channels _sometimes_ (but not always!) generate a
weird channelName, with the name itself being repeated and verification
mark, if present also included as the word "Verified".  In such cases
both the duplicate names and the word "Verified" are separated by
multiple line breaks, which causes the list item for that playlist to
stretch vertically.

Address this by removing whitespace from the channel name both on
loading JSON from storage and on extracting it from HTML before saving
to storage.